refactor(server): move database runtime into package
